+++ src/server/utils/request-errors/request-errors.ts
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2017-2022 Allegro.pl
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export class AccessDeniedError extends Error {
+  code = 403;
+}
+
+export class InvalidRequestError extends Error {
+  code = 400;
+}
+
+export function isAccessDeniedError(e: Error): e is AccessDeniedError {
+  return e instanceof AccessDeniedError;
+}
+
+export function isInvalidRequestError(e: Error): e is InvalidRequestError {
+  return e instanceof InvalidRequestError;
+}
